feat (DPLAN-11799) adjust command to have abstract parent class where shared code is there
